Add unit tests for blockchain node
- Pass genesis blocks to node's constructor explicitly
- Don't include extra MONETARY_FEE for data transactions
- Add basic unit tests for NodeService
- Allow to specify the total number of coins in bootstrap utility
- Use usize instead of u64 for DATA_UNIT
- Improve debug output in Broker
- Remove NodeService::outbox (an artifact)

See #190

+++ blockchain/src/output.rs
@@ -305,9 +305,9 @@ impl DataOutput {
         Ok((delta, gamma, data))
     }
 
-    pub fn data_size(&self) -> u64 {
+    pub fn data_size(&self) -> usize {
         assert!(self.payload.ctxt.len() > DATA_PAYLOAD_LEN);
-        (self.payload.ctxt.len() - DATA_PAYLOAD_LEN) as u64
+        (self.payload.ctxt.len() - DATA_PAYLOAD_LEN)
     }
 }
